remove use of global query var
I started out using the global query var but realized that I should conform
to code style by keeping it in a parameter.

on line 353, I'm not sure why the line was
(1) ... requests.get(url + query) ...
when I believe it should be
(2) ... requests.get(url) ...

Here is the difference:
(1) http://stackoverflow.com/questions/1789945/how-to-check-whether-a-string-contains-a-substring-in-javascript/1789952#1789952foo
(2) http://stackoverflow.com/questions/1789945/how-to-check-whether-a-string-contains-a-substring-in-javascript/1789952#1789952

I believe (2) is correct.