scrolling up and down works in Answers Page
Scrolling up and down in the answer page when the answer was too long
did not work when the Answer ListBox was contained within a Pile
widget. The only way to make scrolling work when there is not enough
room to display the entire answer is to treat the ListBox
as a Box widget, not a flow widget, by placing it inside a Frame.