No need to declare data_files.
We're already declaring `include_package_data=True` and `recursive-include socli *` in MANIFEST.in. This ensures that
the user_agent.txt file will be available in the package installation
path and thus will be accessible from the package's code.

Location resolved for data_files is unreliable, so let's just get rid of
it since we don't need it.